Move ClassLoaderObjectInputStream.java to :messaging
