+++ platforms/core-execution/worker-processes/src/main/java/org/gradle/process/internal/worker/messaging/WorkerConfigSerializer.java
@@ -20,11 +20,11 @@ import org.gradle.api.Action;
 import org.gradle.api.GradleException;
 import org.gradle.api.JavaVersion;
 import org.gradle.api.logging.LogLevel;
-import org.gradle.internal.io.ClassLoaderObjectInputStream;
-import org.gradle.internal.io.ClassLoaderObjectInputStream.UnsupportedClassVersionErrorWithJavaVersion;
 import org.gradle.internal.nativeintegration.services.NativeServices.NativeServicesMode;
 import org.gradle.internal.remote.internal.inet.MultiChoiceAddress;
 import org.gradle.internal.remote.internal.inet.MultiChoiceAddressSerializer;
+import org.gradle.internal.serialize.ClassLoaderObjectInputStream;
+import org.gradle.internal.serialize.ClassLoaderObjectInputStream.UnsupportedClassVersionErrorWithJavaVersion;
 import org.gradle.internal.serialize.Decoder;
 import org.gradle.internal.serialize.Encoder;
 import org.gradle.internal.serialize.Serializer;
